factor out version logic into own private module
